Going back into testing.
Need to find out why a max/min of 1 bugs out.
Need to find out why the wheel doesn't fully stop (suspect speed algorithm's time interval is too small for such low speeds)
Need to find out how to dynamically calculate setpoints.

+++ Digital/CS301_Class.cydsn/control/pid.c
@@ -68,7 +68,7 @@ void pid_compute(PIDData* data) {
         data->output = data->output_min;
     }
     else {
-        data->output = (int32_t)output;
+        data->output = output;
     }
 }
